fixup! fix(core): Access injected parent values using SelfSkip
